Introduce QuicConnectionIdAddressGenerator and remove QuicConnectionSigner
Motivation:

While it is valid to use some sort of signing to generate a new connection id its not really needed (by the spec). We should be more unspecific here and also be consistent with the client side.

Modifications:

- Introduce QuicConnectionIdAddressGenerator and expose a default implementation via QuichConnectionIdAddress that uses a SecureRandom
- Split QuicCodecBuilder into QuicClientCodecBuilder and QuicServerCodecBuilder to ensure we only enforce specific config where its needed
- Update the examples for the new API
- Add some javadocs to the builders

Result:

Be more flexible in how the connection id is generated