Make the usage of QuicConnectionAddress optional
Motivation:

Most people should not need to worry about generating a QuicConnectionAddress at all and should just use the default implementation that we provide. Beside this it should be easier to use a connected Channel as transport layer.

Modification:

- Remove constructor and static method that takes InetSocketAddress as argument in QuicConnectionAddress.
- Add methods to QuicChannelBootstrap that allow to specify the remoteAddress (InetSocketAddress) and the connectionAddress(QuicConnectionAddress)
- If the remoteAddress is not configured try to obtain it from the parent Channel and so make it easier to use a connected Channel as transport
- if the connectionAddress is not configured generated a new random one
- Adjust tests / examples for new api

Result:

Less boilerplate code for most users and also more flexibility