Call Quic.ensureAvailability() before trying to use it and so fail fast
Motivation:

We should call Quic.ensureAvailability() before trying to use quiche. This will ensure we fail fast and make it easier for users to detect whats wrong. Also we missed to catch an exception in the static block in Quic which would have lead to problems when use the code on platforms that we not support atm.

Modifications:

- Add Quic.ensureAvailability() calls
- Add try { ... } catch { ... } in Quiche

Result:

Fail fast on "safe" on platforms that we not support