Route stream creations through the pipeline
Motivation:

Creating streams is generally speaking some sort of establishing a connection. So it kind of make sense to re-use the connect(...) abstraction internally. This will also allow us to intercept these if needed and fail fast (like for example this will be useful in HTTP/3 when a go-away was already received

Modifications:

Let createStream(...) flow through the ChannelPipeline of QuicChannel

Result:

Be able to intercept stream creation attempts.