Only allocate with native order if we really need too
Motivation:

During HTTP/3 profiling it showed up that we spend about 9% of all allocations because we use internalNioBuffer(...) on a SwappedByteBuf. This is not needed for this use-case as we only need to allocate with native order if we read out primitive values like int / long etc.

Modifications:

Only allocate with native order if its really needed.

Result:

Less allocations