Change shutdownOutput() semantics again to do a "clean" shutdown
Motivation:

We recently changes all shutdownOutput() methods to do a shutdown with a RESET_FRAME. This was not a good call and made things harder for users. Let's change it back to only do a reset when shutdownOutput(...) is caled with an error code.

Modifications:

Do a "clean" shutdown aka FIN when shutdownOutput() is called without an error code

Result:

Easier to use again