Make it more flexible in terms of supporting GSO
Motivation:

At the moment we hardcode the support for GSO for epoll. This is not really flexible enough as we may be able to support it for other transport implementations (like io uring).

Modifications:

Introduce a new API that can be used to plugin GSO support

Result:

Remove dependency on epoll and make the GSO support pluggable