Allow to override the repository url for quiche and boringssl
Motivation:

Sometimes a user may want to use a custom fork of either quiche or boringssl but for this the pom.xml needs to be adjusted. We should just allow to override the repository urls.

Modifications:

Allow to override the repository urls by using -D ...

Result:

Easier to use own forks of quiche or boringssl