Stream might get into a bad state when the reader is slow.
Motivation:

It is possible for a QuicheQuicStreamChannel to get into a bad state when the following
events happen:
  - stream becomes writable.
  - connection receives data and QuicheQuicChannel.handleWritableStreams() is called.
  - QuicheQuicStreamChannel.writable() is called with a positive value.
  - QuicStreamChannelUnsafe.writeQueued() causes the capability to get to 0 and the
    queue becomes non-empty.
  - At the end of QuicheQuicStreamChannel.writable(), the stream is set to writable.

From there on, QuicheQuicStreamChannel.write() appends data to the queue since it
is not empty.

Modification:

QuicheQuicStreamChannel.writable() should use the current capability to decide
if the writability has changed.

Result:

Stream is happier when there is a slow reader.