Correctly use jlong and jint in native function definition to fix possible segfault
Motivation:

Due a bug we used long and int in the function definition while it must be jlong and jint. The problem is that jlong may not be the same "size" as long and so using the wrong type may result in truncating of the value. In this case we would use the wrong value as pointer and then crash.

While the crashes were only observed on the CI on windows the reality is that it can happen basically everywhere. Long story short: the 'j' matters ...

Modifications:

Use jlong and jint

Result:

Fixes native crashes