Add methods to QuicChannel that allows to query if a connection was timed out
Motivation:

Sometimes we would like to know if a connection was closed because of a timeout. This was not possible.

Modifications:

Add isTimedOut() method which allows to query if the connection was closed because of the idle time out

Result:

Be able to know if a connection is closed because of the idle timeout