Set the number of streams to a value low enough to not deadlock because of the window
Motivation:

We used a value of 1024 which was too big if we couldnt schedule the reads / writes in the right timely manner. This could lead to a deadlock.

Modifications:

Use a lower value which still provide the required testing but not put us in risk of deadlocks during tests

Result:

More stable builds on CI