Correctly calculate the memory address even if ByteBuf might not have a memory adress that is directly accessible.
Motivation:

When we try to obtain a memory address of a ByteBuf we might need to fallback to obain it via the ByteBuffer. When doing so we also need to take the position of the ByteBuffer into account as it might be non-zero.

Modifications:

- Add new static helper methods to Quiche class that can be used to obtain the reader / writer memory address in a consistent / easy manner.
- Use these new methods

Result:

Always use the correct memory address and so don't end up with data-corruption