Try to send packets as fast as possible during connection establishment
Motivation:

We are trying to be smart and batch as many datagram packets as possible per syscall atm. This is good in terms of CPU usage but might also introduce some latency. This latency might be very noticable during connection establishment. Let's try to "accept" connections as fast as possible

Modifications:

When the Channel was not active yet always directly call connectionSendAndFlush(). This will ensure we produce packets and send them as fast as possible in this case

Result:

Faster connection establishment