Don't leak ByteBuffer when sending segments will use the whole send quantum
Motivation:

When the whole send quantum is used we will exit the loop that tries to write datagram packets. This is correct but we still need to ensure we write the previous filled out Buffer as well. Not doing so will result in a memory leak and also in packet loss.

Modifications:

Even if we return early due used send quantum also write the previous created out Buffer.

Result:

No more leak and also no packet loss when send quantum is used.