Add PR builds with unsafe disabled
Motivation:

We recently had a bug which only showed up when the memoryAddress of a direct buffer could not be accessed directly. We didnt catch this as the memory address always was accessible by unsafe. Let's add a PR build job which explicit disable unsafe.

Modifications:

- Add profile which disables unsafe during build
- Add job which builds with unsafe disabled

Result:

Better test coverage