Only close channel when receiving a STOP_SENDING frame
Motivation:

We need to ensure we correctly handle the situation when receiving a STOP_SENDING frame. Due a bug we fired an exception trough the pipeline and closed the channel. This is confusing, let's just close the channel.

Modifications:

- When receiving a STOP_SENDING frame just close the channel and ensure the writes are failed

Result:

Fixes bug reported in https://github.com/netty/netty-incubator-codec-http3/issues/262