Fire SniCompletionEvent as soon as the SNI hostname was selected
Motivation:

We did fire the SniCompletionEvent only once we were done with the handshake which is not really the correct time to do this. We should fire the event as soon as we were able to process the client hello and selected the hostname.

Modifications:

- Fire the event as soon as we select the hostname

Result:

Fire the event at the correct time