Remove usage of QuicheError in public API as its an implementation detail of Quiche
Motivation:

Our usage of quiche is an implementation detail so we should not leak it to the user. We should convert from quiche errors to errors that are defined in the RFC9000.

Modifications:

Remove QuicheError from public API and replace it by QuicTransportError.

Result:

Cleaner API without leaking implementation details