Only create a new QuicChannel when receiving an INITIAL packet
Motivation:

We should only try to create a new QuicChannel when receiving an INITIAL packets. For all the other packets we want to just drop the packet on the floor if there is no existing mapping.

Modifications:

Only create a new QuicChannel on receiving an INITAL packet

Result:

Correctly handle creation of new QuicChannel