Writability event is triggered with wrong capacity #747
Motivation:

There was a regression in fdb1a7 (00.63.Final) where writeQueued() may call updateWritabilityIfNeeded(true)
at the end of the loop, even when no write operations were completed. As a result, the channel capacity is zero.
That has an impact on handlers that intercept the writability events and call Channel.bytesBeforeUnwritable().

Modifications:

Only call `updateWritabilityIfNeeded(true) if 'written` is true.

Result:

Handlers that intercept writability events see the correct capacity.