Correctly detect supported namedGroups during initialization
Motivation:

When init our quic implementation we also try to respect jdk.tls.namedGroups. While doing so we try to detect which group is supported by the underlying BoringSSL version. Unfortunally we did miss-interpret the return value of BoringSSL.SSLContext_set1_groups_list(...) and so assumed a group is not supported even if it is.

Modifications:

Correctly handle return value of BoringSSL.SSLContext_set1_groups_list(...)

Result:

Be able to configure a supported named group via -Djdk.tls.namedGroups=