Fix crash in Quiche when calling isClosed on freed connections
Motivation:
When QuicheQuicConnection is freed, it sets the connection reference to -1. Unfortunately, it's possible to then inspect the native closed state of this connection, which will crash in Quiche.

Modification:
Change the `assert connection != -1` to a real conditional in `isClosed`, to avoid calling into Quiche on freed connections.

Result:
This fixes a JVM crash bug with our QUIC codec.